Linked_crystal/app/requirements.txt
appdirs==1.4.4
build==1.2.2.post1
buildozer==1.5.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Cython==0.29.36
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
exceptiongroup==1.3.0
filelock==3.18.0
filetype==1.2.0
idna==3.10
Jinja2==3.1.6
Kivy==2.3.0
Kivy-Garden==0.1.5
MarkupSafe==3.0.2
numpy==1.21.6
packaging==25.0
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
ptyprocess==0.7.0
git+https://github.com/pyboy-for-android/PyBoy.git@Desktop-2.6.0#egg=pyboy
Pygments==2.19.1
pyproject_hooks==1.2.0
PySDL2==0.9.17
pysdl2-dll==2.32.0
python-for-android==2024.1.21
requests==2.32.3
scikit-build==0.18.1
scikit_build_core==0.11.4
sh==1.14.3
sounddevice==0.5.2
toml==0.10.2
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
virtualenv==20.31.2
websockets==15.0.1
python-dotenv==1.2.1
plyer==2.1.0

